requirements.txt
gspread
oauth2client
onvif_zeep